Address PR review: combinators rename, drop rejection helper
- Rename src/static.ts -> src/combinators.ts (subpath becomes
  @repeaterjs/repeater/combinators); update index/limiters imports and
  regenerate the exports map.
- Drop the custom rejection() test helper in favor of the built-in
  `expect(p).rejects` matcher, which works on bun for native promises
  (the helper was an over-correction; only the .resolves cases needed
  plain-await).

tsc clean; 163/163 on node and bun.

Refs #81
