+++ package.json
@@ -64,16 +64,6 @@
       "import": "./dist/src/limiters.js",
       "require": "./dist/src/limiters.cjs"
     },
-    "./static": {
-      "types": "./dist/src/static.d.ts",
-      "import": "./dist/src/static.js",
-      "require": "./dist/src/static.cjs"
-    },
-    "./static.js": {
-      "types": "./dist/src/static.d.ts",
-      "import": "./dist/src/static.js",
-      "require": "./dist/src/static.cjs"
-    },
     "./timers": {
       "types": "./dist/src/timers.d.ts",
       "import": "./dist/src/timers.js",
@@ -84,6 +74,16 @@
       "import": "./dist/src/timers.js",
       "require": "./dist/src/timers.cjs"
     },
-    "./package.json": "./dist/package.json"
+    "./package.json": "./dist/package.json",
+    "./combinators": {
+      "types": "./dist/src/combinators.d.ts",
+      "import": "./dist/src/combinators.js",
+      "require": "./dist/src/combinators.cjs"
+    },
+    "./combinators.js": {
+      "types": "./dist/src/combinators.d.ts",
+      "import": "./dist/src/combinators.js",
+      "require": "./dist/src/combinators.cjs"
+    }
   }
 }
