+++ src/test/_testutils.ts
@@ -41,21 +41,6 @@ export async function* hangingGen<T = never>(): AsyncIterableIterator<T> {
   yield (Infinity as unknown) as T;
 }
 
-// Awaits a promise expected to reject and returns the rejection reason. Used
-// instead of expect(p).rejects in timing-sensitive tests, because the bun test
-// runner's .resolves/.rejects matchers insert different microtask ticks than a
-// plain await, which shifts when a repeater's floating-rejection propagation
-// lands relative to subsequent next() calls.
-export async function rejection(p: Promise<unknown>): Promise<unknown> {
-  try {
-    await p;
-  } catch (err) {
-    return err;
-  }
-
-  throw new Error("Expected promise to reject");
-}
-
 export function delayPromise<T>(
   wait: number,
   value?: T,
